worker/requirements.txt
python-dotenv>=1.0.0
supabase>=2.0.0
requests>=2.31.0
pdfplumber>=0.11.0
pypdf>=4.0.0
openai>=1.0.0
httpx>=0.27.0
# OCR support (optional — worker degrades gracefully if not installed)
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
